adding unit test for tests
